+++ src/lib/components/form/Select/index.jsx
@@ -80,7 +80,7 @@ export const Select = (props) => {
       > */}
         <select { ...mergeProps("select", props => ({
           ...props,
-          ...mergeQuickProps(["name", "multiple", "disabled", "readOnly", "onBlur", "onFocus"]),
+          ...mergeQuickProps(props, ["name", "multiple", "disabled", "readOnly", "onBlur", "onFocus"]),
           value: currentValue ?? (multiple ? [] : ""),
           onChange: e => {
             handleSelectOnChange(e);